requirements.txt
hokuyolx==0.9.0
numpy
psycopg2-binary==2.9.3
requests
urllib3
pandas